+++ yolo-yacht-search/includes/class-yolo-ys-stripe.php
@@ -81,7 +81,7 @@ class YOLO_YS_Stripe {
             
             // Create Checkout Session
             $session = \Stripe\Checkout\Session::create([
-                'payment_method_types' => ['card'],
+                'automatic_payment_methods' => ['enabled' => true], // v75.19: Enable all payment methods
                 'line_items' => [[
                     'price_data' => [
                         'currency' => strtolower($currency),
@@ -142,7 +142,7 @@ class YOLO_YS_Stripe {
             
             // Create Checkout Session
             $session = \Stripe\Checkout\Session::create([
-                'payment_method_types' => ['card'],
+                'automatic_payment_methods' => ['enabled' => true], // v75.19: Enable all payment methods
                 'line_items' => [[
                     'price_data' => [
                         'currency' => strtolower($currency),